requirements.txt
pandas>=1.3.0
numpy>=1.20.0
openpyxl>=3.0.0
streamlit>=1.10.0
requests>=2.25.0
python-dateutil>=2.8.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn
sentence-transformers 
torch  
sqlalchemy
python-dotenv
passlib
pydantic_settings
bcrypt
fastapi
uvicorn
python-jose
pydantic[email]
python-multipart
openai
tiktoken